requirements.txt
django==1.10.5
djangorestframework==3.5.4
django-filter==1.0.1
markdown==2.6.8